# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EduApp",
        "description": "Documentation Swagger de l'API d'authentification",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/seances": {
            "post": {
                "tags": [
                    "Séances"
                ],
                "summary": "Filtrer les séances par date et heure",
                "description": "Retourne les séances d’une date donnée avec une heure optionnelle",
                "operationId": "6a02fe5c03c1627d1fc9468d1fd1b53e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date_seance"
                                ],
                                "properties": {
                                    "date_seance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-15"
                                    },
                                    "heure_seance": {
                                        "type": "string",
                                        "example": "08:00-10:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Liste des séances",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "selected_date": {
                                            "type": "string",
                                            "example": "2026-01-15"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seances/{id}/etat": {
            "put": {
                "tags": [
                    "Séances"
                ],
                "summary": "Modifier l’état d’une séance",
                "description": "Met à jour l’état d’une séance (0 = non faite, 1 = faite)",
                "operationId": "7e59347d542d3444caf506bf183cf8b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la séance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "etat"
                                ],
                                "properties": {
                                    "etat": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Etat modifié avec succès"
                    },
                    "404": {
                        "description": "Séance introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/voeux": {
            "get": {
                "tags": [
                    "Voeux Enseignement"
                ],
                "summary": "Lister les vœux de l'enseignant connecté",
                "operationId": "6263b1b7fcf848a796abef0739976373",
                "responses": {
                    "200": {
                        "description": "Liste des vœux avec charge et heures restantes"
                    },
                    "404": {
                        "description": "Profil enseignant introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Voeux Enseignement"
                ],
                "summary": "Ajouter un vœu d'enseignement",
                "operationId": "8406c0c4ae9ad883854ae9c68c3cbb8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_jour",
                                    "code_seance"
                                ],
                                "properties": {
                                    "code_jour": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "code_seance": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vœu ajouté avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation ou dépassement de charge"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/voeux/update": {
            "put": {
                "tags": [
                    "Voeux Enseignement"
                ],
                "summary": "Mettre à jour les vœux (remplacement total)",
                "operationId": "c39331d15315d78aa3beb754c9a12851",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voeux": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "code_jour": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "code_seance": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vœux mis à jour"
                    },
                    "400": {
                        "description": "Charge dépassée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/voeux-enseignement/{id}": {
            "delete": {
                "tags": [
                    "Voeux Enseignement"
                ],
                "summary": "Supprimer un vœu",
                "operationId": "7800bc716e464bf5a8672e8fddab3b6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vœu supprimé"
                    },
                    "404": {
                        "description": "Vœu introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}